% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Action-class}
\alias{Action-class}
\title{Class to handle requests form the UI of Biomon.}
\description{
Action
}
\section{Slots}{

\describe{
\item{\code{success}}{A numeric attribute meaning the status of the execution, where 0 is success any other number a message ID.}

\item{\code{msgType}}{A character meaning 'S' success, E' error, 'W' warning or 'I' information.}

\item{\code{msgText}}{A character string with a message.}

\item{\code{anlz}}{An instance of the class Analyz.}
}}

