% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.btnExecute}
\alias{Action.btnExecute}
\alias{Action.btnExecute,Action,Action-method}
\alias{Action.btnExecute,Action-method}
\title{Action.btnExecute}
\usage{
Action.btnExecute(object, analysisFl)

\S4method{Action.btnExecute}{Action}(object, analysisFl)
}
\arguments{
\item{object}{Description.}

\item{analysisFl}{A list with the paths of the CSV files.}
}
\value{
result The return states the number of results reached or zero when something went wrong.
}
\description{
Description.
}
\examples{
obj <- new("Action")
v_analysisFl <- c()
Action.btnExecute(obj, v_analysisFl)
}

