% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.btnExecute}
\alias{Action.btnExecute}
\title{Action.btnExecute}
\usage{
Action.btnExecute(object, anlzFile)
}
\arguments{
\item{object}{Description.}

\item{anlzFile}{The path and file name of the analysis CSV file.}
}
\value{
result The return states the number of results reached or zero when something went wrong.
}
\description{
Description.
}
\examples{
Action.btnExecute()
}

