% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.btnExecuteEcho}
\alias{Action.btnExecuteEcho}
\alias{Action.btnExecuteEcho,Action,Action-method}
\alias{Action.btnExecuteEcho,Action-method}
\title{Action.btnExecuteEcho}
\usage{
Action.btnExecuteEcho(object, string)

\S4method{Action.btnExecuteEcho}{Action}(object, string)
}
\arguments{
\item{object}{Object instance.}

\item{string}{A string to be "echoed".}
}
\value{
result The return is the string sent in the parameter string.
}
\description{
Method for conneciton testing purposes.
}
\examples{
obj <- new("Action")
v_anySTR <- character()
Action.btnExecuteEcho(obj, v_anySTR)
}

