% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.btnExecuteEcho}
\alias{Action.btnExecuteEcho}
\title{Action.btnExecuteEcho}
\usage{
Action.btnExecuteEcho(object, string)
}
\arguments{
\item{object}{Description.}

\item{string}{A string to be "echoed".}
}
\value{
result The return is the string sent in the parameter string.
}
\description{
Description.
}
\examples{
Action.btnExecuteEcho()
}

