% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.getMsgText}
\alias{Action.getMsgText}
\alias{Action.getMsgText,Action,Action-method}
\alias{Action.getMsgText,Action-method}
\title{Action.getMsgText}
\usage{
Action.getMsgText(object)

\S4method{Action.getMsgText}{Action}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
msgText The returned string describes the result of an execution.
}
\description{
Description.
}
\examples{
obj <- new("Action")
Action.getMsgText(obj)
}

