% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.getMsgText}
\alias{Action.getMsgText}
\title{Action.getMsgText}
\usage{
Action.getMsgText(object)
}
\arguments{
\item{object}{Description.}
}
\value{
msgText The returned string describes the result of an execution.
}
\description{
Description.
}
\examples{
Action.getMsgText()
}

