% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.getMsgType}
\alias{Action.getMsgType}
\title{Action.getMsgType}
\usage{
Action.getMsgType(object)
}
\arguments{
\item{object}{Description.}
}
\value{
msgType The returned type defines the type of the execution message, where 'S' is success, 'E' error, 'W' warning and 'I' information.
}
\description{
Description.
}
\examples{
Action.getMsgType()
}

