% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.getSuccess}
\alias{Action.getSuccess}
\title{Action.getSuccess}
\usage{
Action.getSuccess(object)
}
\arguments{
\item{object}{Description.}
}
\value{
success The returned value states the result of an execution, where 0 is success and any other number means that an issue happens.
}
\description{
Description.
}
\examples{
Action.getSuccess()
}

