% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.setMsgText<-}
\alias{Action.setMsgText,Action,Action-method}
\alias{Action.setMsgText<-}
\alias{Action.setMsgText<-,Action-method}
\title{Action.setMsgText}
\usage{
Action.setMsgText(object) <- value

\S4method{Action.setMsgText}{Action}(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{The informed string describes the result of an execution.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Action")
v_msg <- character()
Action.setMsgText(obj) <- v_msg
}

