% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.setMsgText<-}
\alias{Action.setMsgText<-}
\title{Action.setMsgText}
\usage{
Action.setMsgText(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{The informed string describes the result of an execution.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
Action.setMsgText()
}

