% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.setMsgType<-}
\alias{Action.setMsgType,Action,Action-method}
\alias{Action.setMsgType<-}
\alias{Action.setMsgType<-,Action-method}
\title{Action.setMsgType}
\usage{
Action.setMsgType(object) <- value

\S4method{Action.setMsgType}{Action}(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{The infomed type defines the type of the execution message, where 'S' is success, 'E' error, 'W' warning and 'I' information.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Action")
v_msgType <- character()
Action.setMsgType(obj) <- v_msgType
}

