% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.setMsgType<-}
\alias{Action.setMsgType<-}
\title{Action.setMsgType}
\usage{
Action.setMsgType(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{The infomed type defines the type of the execution message, where 'S' is success, 'E' error, 'W' warning and 'I' information.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
Action.setMsgType()
}

