% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.setSuccess<-}
\alias{Action.setSuccess,Action,Action-method}
\alias{Action.setSuccess<-}
\alias{Action.setSuccess<-,Action-method}
\title{Action.setSuccess}
\usage{
Action.setSuccess(object) <- value

\S4method{Action.setSuccess}{Action}(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{The infomed value states the result of an execution, where 0 is success and any other number means that an issue happens.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Action")
v_success <- numeric()
Action.setSuccess(obj) <- v_success
}

