% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Action.setSuccess<-}
\alias{Action.setSuccess<-}
\title{Action.setSuccess}
\usage{
Action.setSuccess(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{The infomed value states the result of an execution, where 0 is success and any other number means that an issue happens.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
Action.setSuccess()
}

