\name{biomonCore-package}
\alias{biomonCore-package}
\alias{biomonCore}
\docType{package}
\title{
  Actions for biomonCore GUI.
  }
\description{ Description.}

\details{
  \tabular{ll}{
    Package: \tab biomonCore\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2014-10-10\cr
    License: \tab GPL (>=2)\cr
    Dependence: \tab methods\cr
    }
  }
\author{
  Rodrigo Buhler: Who to complain to <rodrigo.buhler@ymail.com>
  }

\keyword{ package }

\examples{ example }