% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gbl.mAnalysis}
\alias{gbl.mAnalysis}
\title{Data from the choosen analysis script file.}
\format{\preformatted{ logi [1, 1] NA
}}
\usage{
gbl.mAnalysis
}
\description{
Data from the choosen analysis script file.
}
\keyword{datasets}

