% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gbl.matrices}
\alias{gbl.matrices}
\title{Path to the matrices files folder.}
\format{\preformatted{ chr "/Users/rodrigobuhler/Dev/R/biomonCore/data/matrices/"
}}
\usage{
gbl.matrices
}
\description{
String with the complete path to the matrices file folder.
}
\keyword{datasets}

