% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gblAnalysis}
\alias{gblAnalysis}
\title{Path to the analysis files folder.}
\format{\preformatted{ chr "/Users/rodrigobuhler/Dev/R/biomonCore/data/analysis/"
}}
\usage{
gblAnalysis
}
\description{
String with the complete path to the analysis files folder.
}
\keyword{datasets}

