% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{class}
\name{Global-class}
\alias{Global-class}
\title{Class to set the Global Environment for the application Biomon.}
\arguments{
\item{arg}{A list of arguments to be used by the class initializer.}
}
\description{
Class Global
}
\section{Slots}{

\describe{
\item{\code{arguments}}{A list of arguments informed in time of creating the object.}
}}

