% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.getAnalysis}
\alias{Global.getAnalysis}
\alias{Global.getAnalysis,Global,Global-method}
\alias{Global.getAnalysis,Global-method}
\title{Method Global.getAnalysis}
\usage{
Global.getAnalysis(object)

\S4method{Global.getAnalysis}{Global}(object)
}
\arguments{
\item{object}{Object instance.}
}
\value{
gbl.analysis Path to the analysis files folder.
}
\description{
Method for returning the analysis files folder path.
}
\examples{
obj <- new("Global", "Biomon")
v_AnalysisPth <- Global.getAnalysis(obj)
}

