% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getAnalysis}
\alias{Global.getAnalysis}
\title{Method Global.getAnalysis}
\usage{
Global.getAnalysis(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.analysis Description.
}
\description{
Description.
}
\examples{
Global.getAnalysis()
}

