% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.getmAnalysis}
\alias{Global.getmAnalysis}
\alias{Global.getmAnalysis,Global,Global-method}
\alias{Global.getmAnalysis,Global-method}
\title{Method Global.getmAnalysis}
\usage{
Global.getmAnalysis(object)

\S4method{Global.getmAnalysis}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mAnalysis Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_AnalysisMtx <- Global.getmAnalysis(obj)
}

