% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getmAnalysis}
\alias{Global.getmAnalysis}
\title{Method Global.getmAnalysis}
\usage{
Global.getmAnalysis(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mAnalysis Description.
}
\description{
Description.
}
\examples{
Global.getmAnalysis()
}

