% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.getMatrices}
\alias{Global.getMatrices}
\alias{Global.getMatrices,Global,Global-method}
\alias{Global.getMatrices,Global-method}
\title{Method Global.getMatrices}
\usage{
Global.getMatrices(object)

\S4method{Global.getMatrices}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.matrices Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_MatricesPth <- Global.getMatrices(obj)
}

