% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getMatrices}
\alias{Global.getMatrices}
\title{Method Global.getMatrices}
\usage{
Global.getMatrices(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.matrices Description.
}
\description{
Description.
}
\examples{
Global.getMatrices()
}

