% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Global.getmAttributes}
\alias{Global.getmAttributes}
\alias{Global.getmAttributes,Global,Global-method}
\alias{Global.getmAttributes,Global-method}
\title{Method Global.getmAttributes}
\usage{
Global.getmAttributes(object)

\S4method{Global.getmAttributes}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mAttributes Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_AttMatrix <- Global.getmAttributes(obj)
}

