% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getmAttributes}
\alias{Global.getmAttributes}
\title{Method Global.getmAttributes}
\usage{
Global.getmAttributes(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mAttributes Description.
}
\description{
Description.
}
\examples{
Global.getmAttributes()
}

