% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.getmEnvironment}
\alias{Global.getmEnvironment}
\alias{Global.getmEnvironment,Global,Global-method}
\alias{Global.getmEnvironment,Global-method}
\title{Method Global.getmEnvironment}
\usage{
Global.getmEnvironment(object)

\S4method{Global.getmEnvironment}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mEnvironment Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_EnvMatrix <- Global.getmEnvironment(obj)
}

