% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getmEnvironment}
\alias{Global.getmEnvironment}
\title{Method Global.getmEnvironment}
\usage{
Global.getmEnvironment(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mEnvironment Description.
}
\description{
Description.
}
\examples{
Global.getmEnvironment()
}

