% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Global.getmFactor}
\alias{Global.getmFactor}
\alias{Global.getmFactor,Global,Global-method}
\alias{Global.getmFactor,Global-method}
\title{Method Global.getmFactor}
\usage{
Global.getmFactor(object)

\S4method{Global.getmFactor}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mFactor Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_FctMatrix <- Global.getmFactor(obj)
}

