% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getmFactor}
\alias{Global.getmFactor}
\title{Method Global.getmFactor}
\usage{
Global.getmFactor(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mFactor Description.
}
\description{
Description.
}
\examples{
Global.getmFactor()
}

