% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.getmSpace}
\alias{Global.getmSpace}
\alias{Global.getmSpace,Global,Global-method}
\alias{Global.getmSpace,Global-method}
\title{Method Global.getmSpace}
\usage{
Global.getmSpace(object)

\S4method{Global.getmSpace}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mSpace Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_SpcMatrix <- Global.getmSpace(obj)
}

