% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getmSpace}
\alias{Global.getmSpace}
\title{Method Global.getmSpace}
\usage{
Global.getmSpace(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mSpace Description.
}
\description{
Description.
}
\examples{
Global.getmSpace()
}

