% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.getmSpecies}
\alias{Global.getmSpecies}
\title{Method Global.getmSpecies}
\usage{
Global.getmSpecies(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mSpecies Description.
}
\description{
Description.
}
\examples{
Global.getmSpecies()
}

