% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.setmAnalysis}
\alias{Global.setmAnalysis}
\alias{Global.setmAnalysis,Global,Global-method}
\alias{Global.setmAnalysis,Global-method}
\title{Method Global.setmAnalysis}
\usage{
Global.setmAnalysis(object, path)

\S4method{Global.setmAnalysis}{Global}(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_AnalysisPth <- character()
Global.setmAnalysis(obj, v_AnalysisPth)
}

