% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.setmAnalysis}
\alias{Global.setmAnalysis}
\title{Method Global.setmAnalysis}
\usage{
Global.setmAnalysis(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
Global.setmAnalysis()
}

