% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.setmAttributes}
\alias{Global.setmAttributes}
\alias{Global.setmAttributes,Global,Global-method}
\alias{Global.setmAttributes,Global-method}
\title{Method Global.setmAttributes}
\usage{
Global.setmAttributes(object, path)

\S4method{Global.setmAttributes}{Global}(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_AttributesPth <- character()
Global.setmAttributes(obj, v_AttributesPth)
}

