% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.setmAttributes}
\alias{Global.setmAttributes}
\title{Method Global.setmAttributes}
\usage{
Global.setmAttributes(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
Global.setmAttributes()
}

