% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassGlobal.R
\docType{methods}
\name{Global.setmEnvironment}
\alias{Global.setmEnvironment}
\alias{Global.setmEnvironment,Global,Global-method}
\alias{Global.setmEnvironment,Global-method}
\title{Method Global.setmEnvironment}
\usage{
Global.setmEnvironment(object, path)

\S4method{Global.setmEnvironment}{Global}(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_EnvPth <- character()
Global.setmEnvironment(obj,v_EnvPth)
}

