% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.setmEnvironment}
\alias{Global.setmEnvironment}
\title{Method Global.setmEnvironment}
\usage{
Global.setmEnvironment(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
Global.setmEnvironment()
}

