% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.setmFactor}
\alias{Global.setmFactor}
\title{Method Global.setmFactor}
\usage{
Global.setmFactor(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
Global.setmFactor()
}

