% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.setmSpace}
\alias{Global.setmSpace}
\title{Method Global.setmSpace}
\usage{
Global.setmSpace(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
Global.setmSpace()
}

