% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Global.setmSpecies}
\alias{Global.setmSpecies}
\alias{Global.setmSpecies,Global,Global-method}
\alias{Global.setmSpecies,Global-method}
\title{Method Global.setmSpecies}
\usage{
Global.setmSpecies(object, path)

\S4method{Global.setmSpecies}{Global}(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_SpeciesPth <- character()
Global.setmSpecies(obj, v_SpeciesPth)
}

