% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Global.setmSpecies}
\alias{Global.setmSpecies}
\title{Method Global.setmSpecies}
\usage{
Global.setmSpecies(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
Global.setmSpecies()
}

