% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{class}
\name{Action-class}
\alias{Action-class}
\title{Class to handle requests form the UI of Biomon.}
\description{
Action
}
\section{Slots}{

\describe{
\item{\code{anlz}}{An instance of the class Analyz.}
}}

