% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.btnExecute}
\alias{Action.btnExecute}
\alias{Action.btnExecute,Action,Action-method}
\alias{Action.btnExecute,Action-method}
\title{Action.btnExecute}
\usage{
Action.btnExecute(object, analysisFl)

\S4method{Action.btnExecute}{Action}(object, analysisFl)
}
\arguments{
\item{object}{Object instance.}

\item{analysisFl}{Path and analysis file name.}
}
\value{
result Result of the execution of the last step described on the analysis file.
}
\description{
Description.
}
\examples{
obj <- new("Action")
v_analysisFl <- c()
Action.btnExecute(obj, v_analysisFl)
}

