% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Action.btnExecute}
\alias{Action.btnExecute}
\alias{Action.btnExecute,Action,Action-method}
\alias{Action.btnExecute,Action-method}
\title{Action.btnExecute}
\usage{
Action.btnExecute(object, pathList)

\S4method{Action.btnExecute}{Action}(object, pathList)
}
\arguments{
\item{object}{Description.}

\item{pathList}{A list with the paths of the CSV files.}
}
\value{
result The return states the number of results reached or zero when something went wrong.
}
\description{
Description.
}
\examples{
obj <- new("Action")
v_pthList <- list()
Action.btnExecute(obj, v_pthList)
}

