% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Action.btnExecuteEcho}
\alias{Action.btnExecuteEcho}
\alias{Action.btnExecuteEcho,Action,Action-method}
\alias{Action.btnExecuteEcho,Action-method}
\title{Action.btnExecuteEcho}
\usage{
Action.btnExecuteEcho(object, string)

\S4method{Action.btnExecuteEcho}{Action}(object, string)
}
\arguments{
\item{object}{Description.}

\item{string}{A string to be "echoed".}
}
\value{
result The return is the string sent in the parameter string.
}
\description{
Description.
}
\examples{
obj <- new("Action")
v_anySTR <- character()
Action.btnExecuteEcho(obj, v_anySTR)
}

