% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.getMsgType}
\alias{Action.getMsgType}
\alias{Action.getMsgType,Action,Action-method}
\alias{Action.getMsgType,Action-method}
\title{Action.getMsgType}
\usage{
Action.getMsgType(object)

\S4method{Action.getMsgType}{Action}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
msgType The returned type defines the type of the execution message, where 'S' is success, 'E' error, 'W' warning and 'I' information.
}
\description{
Description.
}
\examples{
obj <- new("Action")
Action.getMsgType(obj)
}

