% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAction.R
\docType{methods}
\name{Action.getSuccess}
\alias{Action.getSuccess}
\alias{Action.getSuccess,Action,Action-method}
\alias{Action.getSuccess,Action-method}
\title{Action.getSuccess}
\usage{
Action.getSuccess(object)

\S4method{Action.getSuccess}{Action}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
success The returned value states the result of an execution, where 0 is success and any other number means that an issue happens.
}
\description{
Description.
}
\examples{
obj <- new("Action")
Action.getSuccess(obj)
}

