\name{biomonCore-package}
\alias{biomonCore-package}
\alias{biomonCore}
\alias{biomon_run}
\docType{package}

\title{
CSV Analysis Script Runner
}
\description{
Reads tha informed CSV file with a set of methods and their parameters and returns the value of the last execution.
}
\usage{
  biomon_run(i_analysis)
}
\arguments{
  \item{i_analysis}{A CSV file with the analysis customizing.}
}
\details{
\tabular{ll}{
Package: \tab biomonCore\cr
Type: \tab Package\cr
Version: \tab 1.51\cr
Date: \tab 2014-07-07\cr
License: \tab GPL (>=2) \cr
}
An overview about the package
}
\author{
Rodrigo BUhler

Maintainer: Who to complain to <rodrigo.buhler@ymail.com>

}
\references{
Other references for background info.
}

\keyword{ 
 Optionally other standard keywords, one per line, from file
 KEYWORDS in the R documentation directory package 
}
\seealso{
Optional links to other man pages, e.g.
/code{/link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
## Simple examples of the most important functions
}
