% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Global-class}
\alias{Global-class}
\title{Class Global}
\arguments{
\item{arg}{A list of arguments to be used by the class initializer.}
}
\description{
The Global class is responsible to set the Global Environment with the expected variables. The analyst will consider what is defined by this class in time of defining the CSV analysis file.
}
\section{Slots}{

\describe{
\item{\code{arguments}}{A list of arguments informed in time of creating the object.}
}}

