% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Global.getAnalysis}
\alias{Global.getAnalysis}
\alias{Global.getAnalysis,Global,Global-method}
\alias{Global.getAnalysis,Global-method}
\title{Method Global.getAnalysis}
\usage{
Global.getAnalysis(object)

\S4method{Global.getAnalysis}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.analysis Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_AnalysisPth <- Global.getAnalysis(obj)
}

