% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Global.getmSpecies}
\alias{Global.getmSpecies}
\alias{Global.getmSpecies,Global,Global-method}
\alias{Global.getmSpecies,Global-method}
\title{Method Global.getmSpecies}
\usage{
Global.getmSpecies(object)

\S4method{Global.getmSpecies}{Global}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
gbl.mSpecies Description.
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_SpsMatrix <- Global.getmSpecies(obj)
}

