% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Global.setmFactor}
\alias{Global.setmFactor}
\alias{Global.setmFactor,Global,Global-method}
\alias{Global.setmFactor,Global-method}
\title{Method Global.setmFactor}
\usage{
Global.setmFactor(object, path)

\S4method{Global.setmFactor}{Global}(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_FactorPth <- character()
Global.setmFactor(obj, v_FactorPth)
}

