% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Global.setmSpace}
\alias{Global.setmSpace}
\alias{Global.setmSpace,Global,Global-method}
\alias{Global.setmSpace,Global-method}
\title{Method Global.setmSpace}
\usage{
Global.setmSpace(object, path)

\S4method{Global.setmSpace}{Global}(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\description{
Description.
}
\examples{
obj <- new("Global", "Biomon")
v_SpacePth <- character()
Global.setmSpace(obj, v_SpacePth)
}

